requirements.txt
kafka-python==2.0.1
py4j==0.10.4
pyspark==2.2.1
